feat: user 의 is_verified 삭제 및 test case 작성
